Created an installer
The flexgui.iss file can be used with Inno Setup 6.0.3, which is an open-source installer generator (see http://www.jrsoftware.org/isinfo.php for more information).

First, you have to do make zip to create the flexgui directory with all the files in it (if you're not Eric Smith, you can unzip the release zip file in the top directory of the Github repository).

Then, the .iss file can either be loaded into InnoSetup and then compiled and executed from there, or it can be compiled from the command line by using the command:

  "%programfiles(x86)%\Inno Setup 6\iscc.exe" d:\path\to\flexgui.iss

The installer executable will be generated in the same directory as the .iss file.

Something similar to the following can be added to the makefile (I didn't test this; I don't have a working Cygwin/MingW build environment at this time):

ISCC ?= $(PROGRAMFILES(X86))/iscc.exe

winstaller: flexgui.zip
	$(ISCC) winstaller/flexgui.iss

NOTE: The installer puts the binary files and the include files in a directory that is only accessible for administrators. The samples are stored in a directory flexgui under your Documents directory (and aren't deleted on uninstall).

I changed the gui.tcl file to use the $HOME directory to store the configuration file; it may be necessary to inspect the other source files for use of the ROOTDIR variable and determine if anything is generated from it that is supposed to be writeable for regular users.

I did some minor testing with a Propeller 1 and an existing Spin application and it worked. I haven't tested with P2 (yet) and I haven't tested any other languages.