Created an installer
To build the installer yourself, do the following:
1. Unzip the flexgui.zip from the latest release in the top directory of the Git repository, or use the following command: make zip
2. Open the WindowsInstaller\flexgui.iss file in Inno Setup, and click the Compile button (or hit Ctrl+F9 on the keyboard).
3. The flexgui directory should now contain a file called FlexGUISetup-x.y.z.exe (where x.y.z. is the version number that corresponds to the version in version.tcl).

Building the installer is not currently part of the build, but adding the following lines to Makefile should make it possible to add it. The reason I didn't do this is that I haven't been able to build the rest of FlexGUI on my system; the installer works with the files as staged for the Zip file that Erik distributes as release.

ISCC ?= $(PROGRAMFILES(X86))/iscc.exe

winstaller: flexgui.zip
	$(ISCC) winstaller/flexgui.iss

The installer must be started by a user with administrative privileges. If the user doesn't have enough access rights, the usual elevation dialog appears to ask for credentials for someone who does have enough access rights.

The installer creates a directory "%ProgramFiles(x86)%\Total Spectrum Software\FlexGUI" and installs all the programs there.  It makes a small change to the gui.tcl file so that the FlexGUI configuration file is stored in the user's home directory instead of the install directory. This is necessary because without it, it would be impossible for users to save the configuration because the install directory is read-only for normal users.

Installing the documentation and sample code is optional. The documentation is stored in the doc directory under the install directory, but the sample code is installed in the common user data directory so that it can be modified by users.

The program can be uninstalled from the Apps & Features dialog in Windows Settings.

For more information about Inno Setup, go to https://jrsoftware.org/isinfo.php. The installer was made with InnoSetup 6.0.3.